Implement Discovery routine within routing
