+++ internal/host/host.go
@@ -3,12 +3,17 @@ package host
 import (
 	"context"
 	"fmt"
+	"log"
 	"net"
 	"os"
+	"sync"
+	"time"
 
 	"github.com/libp2p/go-libp2p"
 	"github.com/libp2p/go-libp2p-core/host"
+	"github.com/libp2p/go-libp2p-core/peer"
 	"github.com/libp2p/go-libp2p-core/routing"
+	discovery "github.com/libp2p/go-libp2p-discovery"
 	dht "github.com/libp2p/go-libp2p-kad-dht"
 	libp2pquic "github.com/libp2p/go-libp2p-quic-transport"
 	secio "github.com/libp2p/go-libp2p-secio"
@@ -31,7 +36,7 @@ func NewHost(ctx context.Context, olc string) (host.Host, string, error) {
 	}
 
 	// @2. Create Libp2p Host
-	var kademliaDHT *dht.IpfsDHT
+	point := "sonr-kademlia+" + olc
 	h, err := libp2p.New(ctx,
 		// Add listening Addresses
 		libp2p.ListenAddrStrings(
@@ -55,8 +60,36 @@ func NewHost(ctx context.Context, olc string) (host.Host, string, error) {
 
 		// Let this host use the DHT to find other hosts
 		libp2p.Routing(func(h host.Host) (routing.PeerRouting, error) {
+			// Create New IDHT
 			idht, err := dht.New(ctx, h)
-			kademliaDHT = idht
+			if err != nil {
+				return nil, err
+			}
+
+			// Connect to bootstrap nodes
+			var wg sync.WaitGroup
+			for _, peerAddr := range dht.DefaultBootstrapPeers {
+				peerinfo, _ := peer.AddrInfoFromP2pAddr(peerAddr)
+				wg.Add(1)
+				go func() {
+					defer wg.Done()
+					if err := h.Connect(ctx, *peerinfo); err != nil {
+						log.Println(err)
+					} else {
+						log.Println("Connection established with bootstrap node:", *peerinfo)
+					}
+				}()
+			}
+			wg.Wait()
+
+			// We use a rendezvous point "meet me here" to announce our location.
+			// This is like telling your friends to meet you at the Eiffel Tower.
+			log.Println("Announcing ourselves...")
+			routingDiscovery := discovery.NewRoutingDiscovery(idht)
+			discovery.Advertise(ctx, routingDiscovery, point, discovery.TTL((time.Second * 3)))
+			log.Println("Successfully announced!")
+			go handleKademliaDiscovery(ctx, h, routingDiscovery, point)
+			log.Println("Waiting for Peers...")
 			return idht, err
 		}),
 		// Let this host use relays and advertise itself on relays if
@@ -67,7 +100,39 @@ func NewHost(ctx context.Context, olc string) (host.Host, string, error) {
 
 	// setup local mDNS discovery
 	// err = startMDNS(ctx, h, olc)
-	err = startRendezvous(ctx, h, kademliaDHT, olc)
 	fmt.Println("MDNS Started")
 	return h, h.ID().String(), err
+}
+
+// ^ Handles Peers that appear on DHT ^
+func handleKademliaDiscovery(ctx context.Context, h host.Host, disc *discovery.RoutingDiscovery, point string) {
+	// Timer checks to dispose of peers
+	peerRefreshTicker := time.NewTicker(time.Second * 3)
+	defer peerRefreshTicker.Stop()
+
+	// Start Routing Discovery
+	for {
+		select {
+		case <-peerRefreshTicker.C:
+			peerChan, err := disc.FindPeers(ctx, point)
+			if err != nil {
+				log.Println("Failed to find peers: ", err)
+				return
+			}
+			for peer := range peerChan {
+				if peer.ID == h.ID() {
+					continue
+				} else {
+					log.Println("Found peer:", peer)
+					err := h.Connect(ctx, peer)
+					if err != nil {
+						log.Println("Error occurred connecting to peer: ", err)
+					}
+				}
+
+			}
+		case <-ctx.Done():
+			return
+		}
+	}
 }
